pyproject.toml
[tool.poetry]
name = "readme-now-playing"
version = "0.2.0"
description = "good looking cards for your github readme showing your currently playing song on spotify"
authors = ["Aditya Gaikwad <g.aditya2048@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.9"
Flask = "^2.3.3"
requests = "^2.31.0"
firebase-admin = "^6.2.0"
python-dotenv = "^1.0.0"
vibrant-python = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
